arm: gic-v3: handle pending SGI when enabling jailhouse
When doing jailhouse enable/disable stress test, meet an issue
"WARN: unknown SGI received 2", because jailhouse arm is only
able to handle SGI_INJECT(0) and SGI_EVENT(1), so that means
means losing IPIs, including 0 and 1.

In order to avoid losing the interrupt, we need to migrating the
pending physical interrupts(note the plural) into the pending queue,
hardware or, if full, software.

Signed-off-by: Peng Fan <peng.fan@nxp.com>