Documentation: ivshmem: Require that doorbell writes act as memory barriers
This avoids that the guest has to be aware of how the doorbell interrupt
is internally sent to the target CPU because to add the corresponding
memory barrier explicitly. The implementation in Jailhouse already
fulfills this new requirement.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
(cherry picked from commit 368ff6b2acb3f24ec5903db328d73f2c9d986601)