core: Add internal event check before hypervisor activation
If setup code triggers any event that is to be handled on all CPUs, this
should be processed right before VMM activation. Adding this will
consume the superfluous and to-be-removed flush_vcpu_caches. More
importantly, it will enable handling of update_cat properly on Intel
x86. As there might be more cases in the future, install this pattern
for all archs right from the beginning.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>