MLK-24456-2: arm-common: gic-v3: solve incompatibility problems in gic_v3_init()
When the inmate cell is in AArch32 execution state, using mmio_read64()
to obtain the value of GICR_TYPER will cause error because mmio_read64()
generates "ldrd" instruction when compiling 32-bit gic_demo.bin, and
"ldrd" belongs to A64 assembly language which is cannot be used in
AArch32. So use mmio_read32() to read 64-bit GICR_ TYPER in twice and
add processing case for reading the higher 32 bits of GICR_ TYPER in
gicv3_handle_redist_access().

In ARMv8-A, AArch64 state supports four levels of affinity. but AArch32
state can only support three levels of affinity. So set bit[31:24] of
mpidr to be 0 for AArch32, without affecting AArch64 because
"MPIDR_AFFINITY_LEVEL" used in AArch32 and AArch64 is different.

Signed-off-by: Alice Guo <alice.guo@nxp.com>