arm: gic-v3: clear SGI GICR active flag
With Linux Kernel 5.15
commit 6abbd6988971a ("irqchip/gic, gic-v3: Make SGIs use handle_percpu_devid_irq()"),
the on_each_cpu IPI_CALL_FUNC interrupt active flag will not be cleared
until interrupt handler finish.

Without Jailhouse hypervisor enabled, everything is ok, but when
enabling jailhouse, HCR_EL2.[FMO | IMO] is set, that means NS-EL1
is actually accessing ICV_DIR_EL1 when eoi_irq after enter_hypervisor
return. It not able to deactive the interrupt that is actually a
phyiscal irq which in active state.

To ARM64, the IPI_CALL_FUNC is using SGI 1 which is same value as
jailhouse SGI_EVENT.

Then the following `jailhouse cell create` will hang the system, because
the previous 'SGI_EVENT' is in active state and not deactivated, so
the current SGI_EVENT issued not able to interrupt the target cpu core.

To resolve this issue, let's clear the active bit of SGI_EVENT and
SGI_INJECT before back to Linux.

Tested on NXP i.MX8MP-EVK

Signed-off-by: Peng Fan <peng.fan@nxp.com>
[Jan: tune inline comment]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>